Search: Fix error when passing double-anchored, single-word arguments to repo:has.description (#39117)
* Fix passing anchored arguments to repo:has.description
Previously, passing a double-anchored regex pattern (e.g. `^foo$`) to repo:has.description caused a SQL error because within `parseDescriptionPattern()` we were ignoring the `exact` return value from `parseIncludePattern()`, which resulted in a malformed SQL query. This fixes that error by adding start and end anchors to each element of `exact` and adding the resulting array as part of a query condition, if `exact` is non-nil.

* Add comment explaining odd decision
`parseIncludePattern()` has given us a list of exact string matches for our pattern, but we then add start and end anchors to each of those matches before passing the resulting list into the db query. This looks weird and unnecessary at first glance, but we do this because using the Postgres regex operator is more efficient than doing something like `lower(description) IN (%s)`. Adding a comment explaining this decision.

* add more detail to comment