search: Add experimental feature to select search query input suggestions with Enter
This commit adds the ability to select search query input suggestions
with Enter. Previous attempts at this didn't work reliably/satisfactory
because both CodeMirror and Monaco would always select the first
suggestion. If suggestions took a while to load and would juuuust appear
when you where pressing Enter to _submit the query_ you would insert the
first suggestion instead.

@codemirror/autocomplete v6.1 introduces the option to _not_ select the
first suggestion by default, avoiding this problem.
However, since this changes how the query input works (users have to
explicitly select an item before it be inserted), I'm putting this
behind a feature flag.
I'm also removing the "Tab" label from the selected suggestion.

Use "Tab" to select still works as before. However I noticed an issue if
multiple suggestions are shown, none is selected and "Tab" is pressed:
The suggestions popover closes, neither is a cursor visible nor is the
next button selected. It looks like `<body>` receives the focus which is
weird. The user can't easily go back to the input, only by clicking into
it. Even worse, the query input still appears to be focused.
I can't replicate this on https://codemirror.net/try/ so maybe it's some
something weird with our setup.
Given that the feature is behind a flag, I think it's still safe to
continue with this.