Switch FormTriggerArea to use CodeMirror query input by default
I'm so sorry that I missed that before the release. Unfotunately this
means that this component still uses the Monaco version by default.

I don't think it warrants a patch release since the input is still
usable and CodeMirror can still be enabled site-wide by setting flag
accordingly, but it's still unfortunate :-/