Add lazy sync support for Rust packages
Previously, we didn't lazily fetch Rust packages because the Rust
package host was not part of the `PublicCodeHosts` list. Adding
Rust packages to this list makes it possible to lazy sync newly released
crates that haven't been tracked before (including if it's the first
version we encounter for this package).