blob: Replace history entry if next selected is close the currently selected line
With this commit a new history entry for selecting a line is only added
if the newly selected line is > 10 lines away from the currently
selected line.

A new entry is always added if a line range is currently selected or
if the new selection is a line range.

A new entry is also always added when pinning the hovercard or closing
the hovercard.