Code ownership: Add codeownership.Job to filter post-search results (#39030)
This PR adds scaffolding for a new filter Job for the code ownership
RFC. 

The job will be used to do post-search filtering based on the new
[`file:has.owner()`](https://github.com/sourcegraph/sourcegraph/pull/38938)
predicate. I've added it quite at the end of the job list (with only a
few more jobs that run afterwards. 

Conceptually, as we do not have added a code ownership information
source yet, this will filter out all results if a code owner is
selected: no files have any owners yet.