search: Show diagnostics for patterns starting/ending with quotes
This commit adds diagnostics for patterns that start or end with a
quotation mark, including a quick fix action to replace the quote, but
only if we are in literal or lucky mode.

If the pattern is completely enclosed in quotes it can also be replaced
with the `content:` filter. The assumption here is that the user used
quotes as a way to "escape" the query.