batches: Add computed state value to DB
This adds a trigger that updates the new `computed_state` value on a changeset in the DB before writes. This ensures it's always up to date.
With this new column, we don't have to reverse-engineer the go logic from the resolver into SQL conditions anymore, thus making multiple places simpler to reason about and reducing the sources of truth for the logic to compute the state. This'll also make filtering extremely fast because we can now index on the state directly.

Most of the diff is actually just fixing bad state in tests that now showed up.