fix race condition in test
This test has a race condition. It can be triggered consistently by
adding a `defer func() { time.Sleep(100 * time.Millisecond) }()` before
the `defer close(synchronizer)`. Unfortunately, I don't think it's
possible to completely remove the race condition without extending the
API of the group, so in this commit I resort to adding a sleep that
makes hitting the race condition vanishingly unlikely.