internal/gitserver: Add comment and rename shard -> repoInfoReq
I spent an unnecessarily high amount of time trying to "fix" this code
for what I thought was not the optimal way of collecting and sending
the API requests, until I had the aha moment and finally understood
why we're doing it this way.

As a result, added a comment to clarify this as well as renamed shard
to repoInfoReq as it doesn't look like that's the right name here and
confused me somewhat more when I was reading this code.