Faster syntax highlighting
I wanted to have a nice separation between the data format we get from
the server and the format CodeMirror uses for decorations (previous
commit).

However for humongous documents this intermediate step / format
conversion results in a lot of GC activity (because we essentially throw
away the data parsed from JSON). Also using CodeMirror's API to retrieve
line information seems to take up a non-trivial amount of time. Given
that only a small portion of the document is rendered, that is a lot of
cost spent upfront.

Delaying doing work until absolutely necessary, seems to improve the
initial render time by 3x-4x.