Move fetching of repo metadata into repo resolution step Currently, we're fetching repo metadata for matches at the API layer, right before returning a streaming search event. This means we're not able to enhance the types in the `result` package by matching against that metadata during result construction, which means we can't compute match ranges for things like repo description, which means we can't do things like highlight them in the UI (which we want, because that would be sick).
This commit moves fetching repo metadata into the `repos.Resolve()` method, which makes that metadata available to each search job to populate as a field on each `result.Match` type. I'm pretty sure this means we end up making more db calls, but the tradeoff is that we are now able to work with repo metadata at the job level.

Theoretically, we could match against metadata in `frontend` but that's not super desirable since it would introduce actual logic into the operation of converting from a `result.Match` to a `streamhttp.EventMatch`. Having repo metadata populated on each `result.Match` means we can expand the scope of what each match type matches against and keep all of that logic contained within the `result` package.