codemirror blob: Integrate with Sourcegraph Extensions
This commit introduces a way to interface with Sourcegraph extensions so
that we can support hovercards and document highlights but also
decorations and other features.

Where possible (and reasonable) I tried to separate Sourcegraph
extension integration logic as much as possible from the other
CodeMirror logic. CodeMirror makes this easy via _facets_. For example
text decorations are provided via a facet which separates the logic of
converting text decorations to CodeMirror decorations from the logic
that fetches text decorations. This allows us to easily support
Sourcegraph extensions as well as the new blame annotations that are not
integrated into the core workflow.

A similar example is the integration for document highlights.

In same cases it didn't make sense to introduce these extension points
and the Sourcegraph extensions components directly listen to CodeMirror
events.

The hovercards implementation isn't complete yet and I left some TODOs
to follup on, but overally it's a usable state.

Note that the integration with React seems to work but I feel like that
introduces some delay in rendering those parts.

Also note that I switched from using `EditorView.editable` to
`EditorState.readOnly`. This allows us to focus the editor and have a
proper selection. That selection is sent to the extensions host when it
changes. This is currently not synced with selected lines and I don't
know what exactly we want to do in that area.