blob: Syntax highlighting for CodeMirror blob view (#39386)
This is the follow up/improved version of #39116

This commit adds support for syntax highlighting from lsif/scip data that is
returned for some languages. Updating the backend to return lsif/scip
data for every language is done in a separate PR (#39264), but this
PR does not depend on it (if treesitter highlighting is not configured
there simply won't be any highlighting).

Syntax highlighting is done by an extension that takes the JSON encoded
scip data and converts it to something CodeMirror can understand.
Decorations are only generated for the lines that are currently
rendered.

I originally converted line/column ranges to document-offset ranges and
used binary search to find the relevant ranges for the currently
rendered lines. However the overhead of doing the line/column -> offset
conversion was noticeable for large documents, but especially unnecessary
for those because only a small subset of lines would be visible.

Then I changed to an approach that would use the data sent from the
server as is (to avoid creating additional objects in memory and GC-ing
the JSON decoded data) and only add a simple line index. The conversion
from line/column to document offset is now delayed until the moment the
decorations are created. This only works under the assumption that the
server sends back the ranges order and without overlap.

Some of the changes I made (e.g. exporting the replaceValue function)
are not relevant anymore for the final version, but I'll leave them in
for completeness.

Other auxiliary changes in this commit:

Changed the base CodeMirror hook to allow for "manually" dispatching a
transaction to update the value. Without this we would trigger at
least two transactions when loading a file: One for updating the
document and one for updating the syntax highlighting.
Now we can do both in a single transaction.
Added a folder for CodeMirror blob extensions and moved the line numbers
extension there as well.