codemirror blob: First stab at hover cards
This is a first implementation of hovercards. It only renders content,
no actions and it doesn't support "pinning" cards. I tried to
approximate the current hovercard as much as possible but my main goal
was to get it working, not to make it look identical.
I gave the hovercars a bit more "space" but I still don't like how they
look like with few content.

I also add support for document highlights because this will also
hihlight the token underneath the cursor. Unlike the current blob view
the token underneath the cursor isn't rendered in a different color (for
the time being).

I tested this with precise code intel information, I don't know how well
it works for search based information.