internal/gitserver: Improvement  to RepoInfo method (#39358)
This PR makes two functional improvements to the RepoInfo client
method and a semantic one.

The functional improvements are:

1. Limit length of map to max number of shards

For reasons unknown now, we were making an unnecesarily high number of
requests since the `shards` map being created as a factor of number of
repos. But it should only ever be the total number of shards of
gitserver currently running.

2. Do not hide response body for failed HTTP reqs

Currently, if the API server returns anythign but a 200 OK, we return
the HTTP status, but swallow the HTTP body. This will contain
important information that will be helpful to understand the error,
which is not possible as of now.

The semantic improvement is:

1. Add comment and rename shard -> repoInfoReq

I spent an unnecessarily high amount of time trying to "fix" this code
for what I thought was not the optimal way of collecting and sending
the API requests, until I had the aha moment and finally understood
why we're doing it this way.

As a result, added a comment to clarify this as well as renamed shard
to repoInfoReq as it doesn't look like that's the right name here and
confused me somewhat more when I was reading this code.