pass paths through all file diff resolver paths
When I implemented filtering `fileDiffs` to certain paths, I missed that
the compute function was called from each of the different resolver
methods and that paths would need to be propagated through each of
these. This meant things like the `rawDiff` node would not return
results that respected the passed-in paths.