core workflow: First pass on inlining symbols in the file tree (#39295)
This builts on top of Rok's symbol tag work.

It's a very rudementary first pass to inline symbols. I basically
duplicated the logic for the symbols sidebar with some changes:

- No pagination, I'm requesting a high number of symbols from the server
  and hope that this will suffice to load all symbols.
- Only show top level symbols. I don't know how reliable the
  `containerName` data is but if its set I assume it's not a top level
  symbol.

I had to thread some additional values through to the File component to
be able to fetch symbols (I guess another alternative would be to also
ask for symbol information when we fetch the file information in the
first place, but that feels very wasteful).

It takes a moment to load the data, but I found it too distracting to
immediately show a loading indicator. I added some logic to only show it
after 800ms. If there are no symbols or there is an error a
corresponding message will be shown.

A big drawback of this first solution is that it will show "No symbols
found." for files that _cannot_ have symbol information, such as
Markdown files. This feels weird, but I don't know of a good way to
handle that.