gitserver: Reduce to one ArchiveReader method (#39654)
Before, we had client.Archive and ArchiveReader. The only difference was
that ArchiveReader took a sub-repo permissions checker but then called
Archive. Instead, we now just remove Archive and in the two places we
were still using it pass in a nil checker. This is safe as it replicates
the behaviour from before.