search: clean up nits around description match ranges (#39719)
* remove unneeded testing.Short() check
This check is handled in dbtest.NewDB()

* remove unneeded err != nil check
tr.SetError() will no-op if err is nil so no need for this check

* remove unneeded test from jobutil/repos_test.go
This test is not needed because that scenario will never happen. Every repo being searched by RepoSearchJob will match all description patterns that were ANDed together in the original query.