cm file view: Add max width to hovercard and fix(?) positioning problem
This commit adds a max width to the hovercard. Currently a hovercard
might use the full window width if necessary, which is bad on large
screens.

This also adds a callback function to the component to inform the parent
that it was rendered. I hope this fixes the positining issues mentioned
in https://github.com/sourcegraph/sourcegraph/pull/39482#pullrequestreview-1053177608
but I wasn't able to reliable reproduce the issue so I can't verify it.