gitserver: Restore repo pinning functionality (#39871)
When we [switched to this helper method](https://github.com/sourcegraph/sourcegraph/pull/34362/files), we lost repo pinning. Since we have it in the settings, I'd rather restore it or completely remove it, but not keep it broken. Who knows who relies on it. Currently, pinned repos would most likely just be unreachable.