oneclickexport: add external_services table DB query processor
This commit introduces a new type of processors, which perform DB queries to arbitrary tables of `frontend` database.

DB queries are grouped in one array of `ExportRequest` as an alternative to previous "each export part is a separate JSON property" approach.

Test plan: new unit test is added. When the REST endpoint will be in place, manual tests will also be performed.