codemirror file view: Add search widget (#39884)
This adds CodeMirror's default search functionality. This works now
because the file view is focusable. When focused (e.g. by clicking into
it), meta+f will show CodeMirror's search input.

This currently uses it's own default style but we might be able to
either customize it via CSS or we can completely render our own UI (the
search extension allows to do that).