Extension deprecation: Change enableLegacyExtensions flag to default to true (#40081)
This PR reverts #39661

Unfortunately fixing the typo revealed a mistake in the JSON struct
parsing logic. Since `ExperimentalFeatures` is defined as a Go `struct`,
its boolean members will always default to `false`. It appears that the
JSON parsing logic is leaving the default value when no property is
defined in the config file and thus `enableLegacyExtensions` is set to
`false` potentially disabling the extensions (even though this should
only be the behavior after 4.0)