internal/gitservedr/protocol: Set omitempty in RepoUpdateResponse
When the /repo-update endpoint clones a repo, the response returned is
an empty struct with zero-values encoded into the JSON string. Instead
we can omit the  fields completely.

Additionally, reword the docstring of the Error attribute of this
struct for clarity.

Also remove outdated comment about using RepoInfoResponse. There is no
other reference to this mythical struct apart from the one in this
comment.