codemirror blob view: Add support for reference panel (and fix some things) (#40131)
This commit adds support for using the blob view in the reference
panel. Besides handling the reference panel specific props
(`disableStatusBar`, `disableDecorations`), I also had to make some
changes to make selecting lines and scrolling them into view work
properly.

The biggest change is which selection information is sent to the
extension host. Instead of sending the editor's cursor position, I'm now
sending the selected lines, which is what the old blob view does. This
also fixes the issue with showing line decorations for selected
lines (instead of just line where the current cursor is).

Having a real cursor is still enabled because we need that for search
(Meta+f) to work.