codemirror file view: Add "click to go to definition" and other fixes (#40232)
This commit

- adds support for "click to go to definition"
- hides the caret
- fixes hover token/active range highlighting
- hide close button for non-pinned hovercards

**Click to go to definition**

Normally the WebHoverOverlay expects to be passed an element to which it
can attach a click event handler to provide this functionality. With
CodeMirror we normally to not operate on the DOM level.
So I extended WebHoverOverlay to accept an alternative source for click
events. When the CodeMirror extenion creates a hovercard, it also
registers a callback that should be called if a click happens within the
active range (see also the third point below).

**Hide caret**

The behavior of "click to go to definition" seems to be at odds with
having a real caret. Having a crate is actually a side effect of making
the blob view focusable and thus being able to trigger CodeMirror's
search functionality.
While a caret might be useful if/when we make the blob view more "IDE
like", for the time being we don't have an actual use for it.
This commit hides the caret (and disables some basic cursor movement
keys) to achieve parity with the current blob view.

**Fix highlighting for active ranges**

Until now only the word underneath the cursor was highlighted, as
determined by CodeMirror. However, the requested hover information also
includes a range to which this information applies. This range should be
used for highlighting and for click to jump to definition behavior.
A couple of places that work with the hover ranges needed to be updated
to account for that.

**Hide close button for non-pinned hovercards**

The close button wasn't functional for non-pinned hovercards and it's
arguable whether it should exist at all. This commit only shows the
close button when the hovercard is pinned.