codemirror file view: Make view non-editable
The original reason for making the view editable (but readonly) was to
make it focusable so that Mod-f would trigger CodeMirror's search
functionality.

The side effect of this was that the view would show a real caret, which
is not desirable at this point. Furthermore it changes how the view can
be scrolled via the keyboard.

PR #40232 tried to solve the caret problem by simply hiding it, but the
keyboard scroll issues remained.

This PR makes the view non-editable (which solves caret and keyboard
scroll issues) but keeps it focusable (to keep Mod-f working) by adding
`tabindex=0` to its content element.

Without the CSS changes it wouldn't be apparent that/when the blob view has
focus.