codemirror file view: Cleanup/simplify sourcegraph extensions integration (#40213)
This commit simplifies the Sourcegraph extensions code. The different
integrations require access to the `context` object, which provides the
extensions API object. So far this was made available via a state field.
However since most of the integrations operate on observables it seems
much simplier and cleaner to pass that obserable to the integrations
when they are created.