codemirror file view: Fix non-extension decorations
While working on depecrating the extension API @philipp-spiess
discovered that blame view data wasn't shown properly when it was
enabled.

The reason for this was that the view plugin used to generate the
CodeMirror decorations didn't do that when the plugin was instantiated.
I can only assume that with extensions enabled it "worked" because the
plugin was instantiated prior to receiving the git blame data. But when
extensions are disabled the plugin is only created once the data is
provided.

This also fixes a subtle bug with how the gutter renders selected lines.