database: ListIndexableReposOptions can take CloneStatus
Instead of a boolean field OnlyUncloned, we can make this take the
underlying enum. I intend to use this to only return cloned repos in
another call site.

This should not change any behaviour. CloneStatusUnknown is the empty
string, which is the default value. That should have the same behaviour
as when OnlyUncloned was false (the default value).

Test Plan: CI

plz-review-url: https://plz.review/review/9961