database: only tell zoekt to index cloned repos
We have a rather high error rate on sourcegraph.com around failing to
index cloning repositories. Currently we have code which ignores
failures due to uncloned repos and didn't handle cloning repos. However,
instead of updating that call site, we can now just tell zoekt to only
start indexing a repo once it is cloned. Historically this wasn't done
since when this was implemented we didn't have cloned state in the DB.

Test Plan: add a repo to the local dev env and see it appear in the
index. This will also be caught by our gqltests in CI.

plz-review-url: https://plz.review/review/9962