Allow users to list repositories currently being cloned
This introduced a new filter to the GraphQL: `cloneStatus`. If set, it
returns only repositories that have the given clone status.

With that we can reduce the number of booleans we use when interacting
with the API: instead of querying for `{cloned: true, notCloned: false}`
(which we did even though the `notCloned` was useless) we can query for
`{cloneStatus: "CLONED"}`.

And now we can also query for `{cloneStatus: "CLONING"}`.

This also changes how the site-admin UI works:
* previously: listing "not cloned" would include "cloning" repositories
* now: listing "not cloned" only lists repositories that are "not cloned".